Split TF and bridge utils into separate files
