+++ ngraph_bridge/utils.cc
@@ -19,16 +19,6 @@
 #include <iostream>
 #include <sstream>
 
-#include "tensorflow/core/common_runtime/dma_helper.h"
-#include "tensorflow/core/common_runtime/optimization_registry.h"
-#include "tensorflow/core/framework/attr_value_util.h"
-#include "tensorflow/core/framework/graph.pb.h"
-#include "tensorflow/core/framework/node_def_util.h"
-#include "tensorflow/core/graph/graph.h"
-#include "tensorflow/core/lib/strings/str_util.h"
-#include "tensorflow/core/platform/default/logging.h"
-#include "tensorflow/core/platform/protobuf.h"
-
 #include "utils.h"
 #include "version.h"
 
@@ -36,169 +26,7 @@ using namespace std;
 
 namespace tensorflow {
 namespace ngraph_bridge {
-namespace util {
-
-template <typename T>
-static void TensorDataToStream(std::ostream& ostream, int64 n_elements,
-                               const char* data) {
-  const T* data_T = reinterpret_cast<const T*>(data);
-  for (size_t i = 0; i < n_elements; i++) {
-    ostream << data_T[i] << ",";
-  }
-}
-
-Status TensorToStream(std::ostream& ostream, const Tensor& tensor) {
-  const char* data = tensor.tensor_data().data();
-  int64 n_elements = tensor.NumElements();
-  switch (tensor.dtype()) {
-    case DT_HALF:
-      TensorDataToStream<Eigen::half>(ostream, n_elements, data);
-      break;
-    case DT_FLOAT:
-      TensorDataToStream<float>(ostream, n_elements, data);
-      break;
-    case DT_DOUBLE:
-      TensorDataToStream<double>(ostream, n_elements, data);
-      break;
-    case DT_UINT32:
-      TensorDataToStream<uint32>(ostream, n_elements, data);
-      break;
-    case DT_INT32:
-      TensorDataToStream<int32>(ostream, n_elements, data);
-      break;
-    case DT_UINT8:
-    case DT_QUINT8:
-      TensorDataToStream<uint8>(ostream, n_elements, data);
-      break;
-    case DT_UINT16:
-    case DT_QUINT16:
-      TensorDataToStream<uint16>(ostream, n_elements, data);
-      break;
-    case DT_INT8:
-    case DT_QINT8:
-      TensorDataToStream<int8>(ostream, n_elements, data);
-      break;
-    case DT_INT16:
-    case DT_QINT16:
-      TensorDataToStream<int16>(ostream, n_elements, data);
-      break;
-    case DT_UINT64:
-      TensorDataToStream<uint64>(ostream, n_elements, data);
-      break;
-    case DT_INT64:
-      TensorDataToStream<int64>(ostream, n_elements, data);
-      break;
-    case DT_BOOL:
-      TensorDataToStream<bool>(ostream, n_elements, data);
-      break;
-    case DT_BFLOAT16:
-      return errors::Internal(
-          "TensorToStream got data type bfloat16. No compatible standard C++ "
-          "data type.");
-      break;
-    default:
-      return errors::Internal("TensorToStream got unsupported data type ",
-                              DataType_Name(tensor.dtype()));
-      break;
-  }
-  return Status::OK();
-}
-
-Status TFDataTypeToNGraphElementType(DataType tf_dt,
-                                     ngraph::element::Type* ng_et) {
-  switch (tf_dt) {
-    case DataType::DT_FLOAT:
-      *ng_et = ngraph::element::f32;
-      break;
-    case DataType::DT_DOUBLE:
-      *ng_et = ngraph::element::f64;
-      break;
-    case DataType::DT_INT32:
-      *ng_et = ngraph::element::i32;
-      break;
-    case DataType::DT_UINT8:
-      *ng_et = ngraph::element::u8;
-      break;
-    case DataType::DT_INT8:
-      *ng_et = ngraph::element::i8;
-      break;
-    case DataType::DT_UINT16:
-      *ng_et = ngraph::element::u16;
-      break;
-    case DataType::DT_INT64:
-      *ng_et = ngraph::element::i64;
-      break;
-    case DataType::DT_UINT32:
-      *ng_et = ngraph::element::u32;
-      break;
-    case DataType::DT_UINT64:
-      *ng_et = ngraph::element::u64;
-      break;
-    case DataType::DT_BOOL:
-      *ng_et = ngraph::element::boolean;
-      break;
-    case DataType::DT_QINT8:
-      *ng_et = ngraph::element::i8;
-      break;
-    case DataType::DT_QUINT8:
-      *ng_et = ngraph::element::u8;
-      break;
-    case DataType::DT_QINT32:
-      *ng_et = ngraph::element::i32;
-      break;
-    case DataType::DT_BFLOAT16:
-      *ng_et = ngraph::element::bf16;
-      break;
-    case DataType::DT_HALF:
-      *ng_et = ngraph::element::f16;
-      break;
-    default:
-      return errors::Unimplemented("Unsupported TensorFlow data type: ",
-                                   DataType_Name(tf_dt));
-  }
-  return Status::OK();
-}
-
-Status TFTensorShapeToNGraphShape(const TensorShape& tf_shape,
-                                  ngraph::Shape* ng_shape) {
-  for (int i = 0; i < tf_shape.dims(); i++) {
-    if (tf_shape.dim_size(i) < 0) {
-      return errors::InvalidArgument(
-          "TensorFlow shape has a negative dimension size");
-    }
-  }
-
-  *ng_shape = ngraph::Shape(tf_shape.dims());
-  for (int i = 0; i < tf_shape.dims(); i++) {
-    (*ng_shape)[i] = tf_shape.dim_size(i);
-  }
-
-  return Status::OK();
-}
-
-void PrintNodeHistogram(const std::unordered_map<string, int>& histogram,
-                        bool sorted) {
-  int histogram_size = histogram.size();
-  if (histogram_size == 0) {
-    std::cout << "None";
-  } else {
-    vector<std::pair<string, int>> vec(begin(histogram), end(histogram));
-    if (sorted) {
-      sort(begin(vec), end(vec),
-           [](const pair<string, int>& a, const pair<string, int>& b) {
-             // descending sort
-             return a.second > b.second;
-           });
-    }
-
-    for (auto node : vec) {
-      bool endelem = node == vec.back();
-      std::cout << " " << node.first << " -> " << node.second
-                << (endelem ? " " : ",");
-    }
-  }
-  std::cout << std::endl;
-}
+namespace utils {
 
 void MemoryProfile(long& vm_usage, long& resident_set) {
   vm_usage = 0;
@@ -223,24 +51,6 @@ void MemoryProfile(long& vm_usage, long& resident_set) {
   }
 }
 
-void DumpTFGraph(tensorflow::Graph* graph, int idx, std::string filename) {
-  if (!DumpAllGraphs()) {
-    return;
-  }
-
-  std::stringstream ss;
-  ss << filename << "_" << std::setfill('0') << std::setw(4) << idx;
-  NGRAPH_VLOG(0) << "Dumping TF graph to " << ss.str() + ".pbtxt";
-
-  GraphDef g_def;
-  graph->ToGraphDef(&g_def);
-
-  string graph_pb_str;
-  protobuf::TextFormat::PrintToString(g_def, &graph_pb_str);
-  std::ofstream ostrm_out(ss.str() + ".pbtxt", std::ios_base::trunc);
-  ostrm_out << graph_pb_str;
-}
-
 void DumpNGGraph(std::shared_ptr<ngraph::Function> function,
                  const string filename) {
   if (!DumpAllGraphs()) {
@@ -257,15 +67,6 @@ void DumpNGGraph(std::shared_ptr<ngraph::Function> function,
 
 bool DumpAllGraphs() { return GetEnv("NGRAPH_TF_DUMP_GRAPHS") == "1"; }
 
-bool IsAlreadyProcessed(Graph* g) {
-  // TODO: place a dummy node as a marker
-  // Current method may fail when graph has no encapsulates after first pass
-  for (Node* node : g->nodes()) {
-    if (node->type_string() == "_nGraphEncapsulate") return true;
-  }
-  return false;
-}
-
 string GetEnv(const char* env) {
   const char* val = std::getenv(env);
   if (val == nullptr) {
@@ -277,6 +78,6 @@ string GetEnv(const char* env) {
 
 void SetEnv(const char* env, const char* val) { setenv(env, val, 1); }
 
-}  // namespace util
+}  // namespace utils
 }  // namespace ngraph_bridge
 }  // namespace tensorflow
